UIAutotest/customerUITest/requirements.txt
allure-pytest==2.8.12
allure-python-commons==2.8.12
fake-useragent==0.1.11
Faker==2.0.1
pluggy==0.13.1
py==1.8.0
pyee==6.0.0
PyMySQL==0.9.3
pyparsing==2.4.1.1
pyppeteer==0.0.25
pyquery==1.4.0
pytest==4.0.0
pytest-forked==1.0.2
pytest-html==1.21.1
pytest-metadata==1.8.0
pytest-repeat==0.8.0
pytest-rerunfailures==7.0
pytest-xdist==1.29.0
python-dateutil==2.8.0
PyYAML==5.1.1
requests==2.22.0
requests-html==0.10.0
selenium==3.141.0
six==1.12.0